+++ diesel/src/pg/types/date_and_time/mod.rs
@@ -111,6 +111,8 @@ impl ToSql<types::Timestamp, Pg> for PgTimestamp {
     }
 }
 
+debug_to_sql!(types::Timestamp, PgTimestamp);
+
 impl FromSql<types::Timestamp, Pg> for PgTimestamp {
     fn from_sql(bytes: Option<&[u8]>) -> Result<Self, Box<Error+Send+Sync>> {
         FromSql::<types::BigInt, Pg>::from_sql(bytes)
@@ -124,6 +126,8 @@ impl ToSql<types::Timestamptz, Pg> for PgTimestamp {
     }
 }
 
+debug_to_sql!(types::Timestamptz, PgTimestamp);
+
 impl FromSql<types::Timestamptz, Pg> for PgTimestamp {
     fn from_sql(bytes: Option<&[u8]>) -> Result<Self, Box<Error+Send+Sync>> {
         FromSql::<types::Timestamp, Pg>::from_sql(bytes)
@@ -135,6 +139,8 @@ impl ToSql<types::Date, Pg> for PgDate {
         ToSql::<types::Integer, Pg>::to_sql(&self.0, out)
     }
 }
+
+debug_to_sql!(types::Date, PgDate);
 
 impl FromSql<types::Date, Pg> for PgDate {
     fn from_sql(bytes: Option<&[u8]>) -> Result<Self, Box<Error+Send+Sync>> {
@@ -148,6 +154,8 @@ impl ToSql<types::Time, Pg> for PgTime {
         ToSql::<types::BigInt, Pg>::to_sql(&self.0, out)
     }
 }
+
+debug_to_sql!(types::Time, PgTime);
 
 impl FromSql<types::Time, Pg> for PgTime {
     fn from_sql(bytes: Option<&[u8]>) -> Result<Self, Box<Error+Send+Sync>> {
@@ -164,6 +172,8 @@ impl ToSql<types::Interval, Pg> for PgInterval {
         Ok(IsNull::No)
     }
 }
+
+debug_to_sql!(types::Interval, PgInterval);
 
 impl FromSql<types::Interval, Pg> for PgInterval {
     fn from_sql(bytes: Option<&[u8]>) -> Result<Self, Box<Error+Send+Sync>> {